move more internal code into detail directory
