+++ include/uos/detail/autoselect.hpp
