+++ include/uos/device/msp430/task_list.hpp
@@ -1,4 +1,4 @@
-#include <uos/task_list.hpp>
+#include <uos/detail/task_list.hpp>
 #include <uos/device/msp430/scheduler.hpp>
 
 namespace uos::dev::msp430 {
